Working on bifurcation code.

+++ src/marginal_probability_density_evolution_monte_carlo.h
@@ -71,8 +71,8 @@ typedef struct {
     UINT32 *runsSecondCluster;
     double *meansFirstCluster;
     double *meansSecondCluster;
-    double timeFirstCluster;
-    double timeSecondCluster;
+    UINT32 timeFirstCluster;
+    UINT32 timeSecondCluster;
     UINT32 *meanPathCluster1;
     UINT32 *meanPathCluster2;
     BOOL *isBifurcated;
